Reorganise test files
