+++ tests/__init__.py
@@ -1,1 +1,1 @@
-"""Tests."""
+"""Tests grouped by parser format."""
