+++ tests/general/__init__.py
@@ -1,0 +1,1 @@
+"""Tests covering multiple brokers or shared features."""
