+++ tests/general/data/test_run_with_example_files_output.txt
@@ -1,5 +1,5 @@
 INFO: No schwab Equity Award JSON file provided
-Parsing tests/test_data/trading212/from_2020-09-11_to_2021-04-02.csv...
+Parsing tests/trading212/data/from_2020-09-11_to_2021-04-02.csv...
 INFO: No sharesight file provided
 INFO: No raw file provided
 INFO: No vanguard file provided
