+++ tests/raw/test_raw.py
@@ -1,24 +1,27 @@
-"""Test Schwab."""
+"""Test raw format support."""
 
 from pathlib import Path
 import subprocess
 
-from .utils import build_cmd
+from tests.utils import build_cmd
 
 
-def test_run_with_schwab_example_files() -> None:
+def test_run_with_raw_files_no_balance_check() -> None:
     """Runs the script and verifies it doesn't fail."""
     cmd = build_cmd(
         "--year",
-        "2023",
-        "--schwab",
-        "tests/test_data/schwab/schwab_transactions-2023.csv",
+        "2022",
+        "--raw",
+        "tests/raw/data/test_data.csv",
+        "--no-balance-check",
     )
     result = subprocess.run(cmd, check=True, capture_output=True)
     stderr_lines = result.stderr.decode().strip().split("\n")
     assert len(stderr_lines) == 1
-    assert stderr_lines[0] == "WARNING: No Schwab Award file provided"
-    expected_file = Path("tests") / "test_data" / "schwab" / "expected_output.txt"
+    assert stderr_lines[0].startswith("WARNING: Bed and breakfasting for META"), (
+        "Unexpected stderr message"
+    )
+    expected_file = Path("tests") / "raw" / "data" / "expected_output.txt"
     expected = expected_file.read_text()
     cmd_str = " ".join([param if param else "''" for param in cmd])
     assert result.stdout.decode("utf-8") == expected, (
