+++ tests/schwab/data/equity_award/schwab_equity_award_v1.json
