+++ tests/schwab/test_schwab_cash_merger.py
@@ -3,7 +3,7 @@
 from pathlib import Path
 import subprocess
 
-from .utils import build_cmd
+from tests.utils import build_cmd
 
 
 def test_run_with_schwab_cash_merger_files() -> None:
@@ -12,7 +12,7 @@ def test_run_with_schwab_cash_merger_files() -> None:
         "--year",
         "2020",
         "--schwab",
-        "tests/test_data/schwab_cash_merger/transactions.csv",
+        "tests/schwab/data/cash_merger/transactions.csv",
     )
     result = subprocess.run(cmd, check=True, capture_output=True)
     stderr_lines = result.stderr.decode().strip().split("\n")
@@ -21,7 +21,7 @@ def test_run_with_schwab_cash_merger_files() -> None:
     assert stderr_lines[0] == "WARNING: No Schwab Award file provided"
     assert stderr_lines[1].startswith("WARNING: Cash Merger support is not complete")
     expected_file = (
-        Path("tests") / "test_data" / "schwab_cash_merger" / "expected_output.txt"
+        Path("tests") / "schwab" / "data" / "cash_merger" / "expected_output.txt"
     )
     expected = expected_file.read_text()
     cmd_str = " ".join([param if param else "''" for param in cmd])
