+++ tests/schwab/test_schwab_equity_award_json.py
@@ -52,7 +52,7 @@ def test_decimal_from_number_or_str_default() -> None:
 def test_schwab_transaction_v1() -> None:
     """Test read_schwab_equity_award_json_transactions() on v1 data."""
     transactions = schwab_equity_award_json.read_schwab_equity_award_json_transactions(
-        "tests/test_data/schwab_equity_award_v1.json"
+        "tests/schwab/data/equity_award/schwab_equity_award_v1.json"
     )
 
     assert transactions[0].date == datetime.date(2022, 4, 25)
@@ -90,7 +90,7 @@ def test_schwab_transaction_v1() -> None:
 def test_schwab_transaction_v2() -> None:
     """Test read_schwab_equity_award_json_transactions() on v2 data."""
     transactions = schwab_equity_award_json.read_schwab_equity_award_json_transactions(
-        "tests/test_data/schwab_equity_award_v2.json"
+        "tests/schwab/data/equity_award/schwab_equity_award_v2.json"
     )
 
     assert transactions[0].date == datetime.date(2023, 4, 25)
@@ -136,7 +136,7 @@ def test_schwab_transaction_v2_rounding() -> None:
     This tests 13 vesting events with 7 shares each, which are then sold.
     """
     transactions = schwab_equity_award_json.read_schwab_equity_award_json_transactions(
-        "tests/test_data/schwab_equity_award_v2_rounding.json"
+        "tests/schwab/data/equity_award/schwab_equity_award_v2_rounding.json"
     )
 
     assert transactions[0].date == datetime.date(2020, 4, 24)
