+++ tests/sharesight/data/inputs/All Trades Report - Test.csv
