+++ tests/trading212/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for Trading212."""
