+++ tests/vanguard/data/expected_output.txt
